Fixed swap page issue
