+++ src/templates/dashboard/index.tsx
@@ -8,6 +8,7 @@ import Transactions from './blocks/Transactions'
 import * as S from './styles'
 import Toast from '../../components/general/Toast'
 import Navbar from './blocks/Navbar'
+import { useTheme ,Theme} from '../../Context/ThemeContext'
 
 export default function Dashboard({ account, blockchainApi }) {
 
@@ -24,6 +25,7 @@ export default function Dashboard({ account, blockchainApi }) {
   const [price, setPrice] = useState<string>('0')
   const [amount, setAmount] = useState<string>('0')
   const [activeIndex, setActiveIndex] = useState(3);
+  const { theme, setTheme } = useTheme();
 
   const removeTransactionsOrder = (id: string) => console.log('remove transaction' + id);
 
@@ -103,7 +105,18 @@ export default function Dashboard({ account, blockchainApi }) {
     setOpenOrders(finalOrders);
   }
 
+  const getInitialTheme = () => {
+    let initialTheme = localStorage.getItem('theme');
+    if (initialTheme) {
+      setTheme(initialTheme === Theme.Light ? Theme.Light : Theme.Dark)
+    } else {
+      initialTheme = Theme.Dark
+    }
+    return initialTheme;
+  }
+
   useEffect(() => {
+    localStorage.setItem('theme', getInitialTheme());
     const webSocketInstance = webSocket;
     fetchMarketData(webSocketInstance)
     fetchOrderBookBids(webSocketInstance)
